producerResponseRate
