--- client/components/ProducerBarGragh.jsx
@@ -1,60 +1,0 @@
-import React, { useState }  from "react";
-// import { Line } from "react-chartjs-2";
-import { useEffect } from "react";
-// import {CategoryScale, Chart} from 'chart.js'; 
-// Chart.register(CategoryScale) 
-
-import {
-  Chart as ChartJS,
-  CategoryScale,
-  LinearScale,
-  BarElement,
-  Title,
-  Tooltip,
-  Legend,
-} from 'chart.js';
-import { Bar } from 'react-chartjs-2';
-
-
-ChartJS.register(
-  CategoryScale,
-  LinearScale,
-  BarElement,
-  Title,
-  Tooltip,
-  Legend
-);
-
-const MakeProducerBarGraph = () => {
-  
-  
-
-  return (
-    <div>
-      <h3 style ={{textAlign : 'center'}}>Total Messages Sent</h3>
-      <div style={{height:"300px", width:"300px"}}>
-      <Bar type='bar' data={ {
-        labels: ['1', '2', '3', '4', '5', '6', '7', '8'],
-        datasets: [
-          {
-            label: 'Total Messages Sent',
-            data:[800, 700, 750, 600, 900, 600],
-            backgroundColor: 'rgba(75, 192, 192, 0.6)',
-            borderColor:'rgba(75, 192, 192, 1)',
-            borderWidth: 4,
-            barPercentage:0.9
-            }
-            ]
-            }} options={{
-              scales: {
-                y: {
-                  beginAtZero: true
-                  }
-                  }}}
-        />
-      </div>
-    </div>
-  );
-};
-
-export default MakeProducerBarGraph;
