+++ client/components/ProducerGraphContainer.jsx
@@ -1,7 +1,9 @@
 import React, { useState } from 'react';
 import { Box, Grid, Typography } from '@material-ui/core';
-import MakeProducerLineGraph from './ProducerLineGraph.jsx';
-import MakeProducerBarGraph from './ProducerBarGragh.jsx';
+import MakeResponseRateGraph from './ProducerResponeRate.jsx';
+import MakeRequestRateGraph from './ProducerRequestRate.jsx';
+import MakeRequestTotalGraph from './ProducerRequestTotal.jsx';
+
 
 
 
@@ -9,20 +11,25 @@ export default function ProducerGraphContainer(){
   return(
     <Box sx={{textAlign : 'center'}}>
     <Typography variant ='h4'>Producer 1</Typography>
+    {/* <div style ={{display: 'flex', flexDirection: 'column', alignItems: 'center'}}> */}
     <Grid
     container
     spacing={4}
-    direction="row"
+    direction="column"
     justifyContent="center"
     alignItems="flex-start"
   >
   <Grid item xs={4}>
-    <MakeProducerLineGraph /> 
+    <MakeResponseRateGraph /> 
   </Grid>
   <Grid item xs={4}>
-    <MakeProducerBarGraph/>
+    <MakeRequestRateGraph/>
+  </Grid>
+  <Grid item xs={4}>
+    <MakeRequestTotalGraph/>
   </Grid>
     </Grid>
+    {/* </div> */}
     </Box>
   )
 }
