+++ client/components/ProducersPage.jsx
@@ -10,14 +10,14 @@ export default function ProducersPage(){
 <div style = {{backgroundImage: 'linear-gradient(to top right, #66a6ff, #89f7fe)'}}>
 <TopBannerContainerProdPage/>
 <Box >
-<Grid container>
-  <Grid item xs={3}>
+{/* <Grid container > */}
+  {/* <Grid item xs={3}>
   <ProducerColumnContainer />  
-  </Grid>
-  <Grid item xs={9}>
+  </Grid> */}
+  <Grid item xs={9} >
   <ProducerGraphContainer />
 </Grid>
-</Grid>
+{/*  </Grid> */}
 </Box>  
 </div>
   )
